JsonNodeUnmarshallContextTesting2 extends JsonNodeUnmarshallContextTesting
